Update username to timeAndUsername & refactor code
Updated where 'username' is used to 'timeAndUsername' to improve clarity - as user input provided now includes time as well. Coordinated class BaseModel and MessageModel to enhance readability and excellence - with the MessageTypes altered from int to byte for performance optimization and replaced the cumbersome constructor in MessageModel by individual setter methods. Revamped ReactionPopupMenuImpl and ReplyPanelImpl to make use of the updated MessageModel. Adjusted EnvVarHandler and CustomUserProperties to reflect the label change from 'username' to 'timeAndUsername'. Reactions are now displayed with a sticker - done in ReactionStickerImpl - derived from the user interaction on a message. Renamed ReactionToSocketDTO to ReactionPanelDTO to better denote its function - transferring ReactionPanel related data. Confirmed all document edits for efficient user interaction behaviour and socket functionality for the main frame was included for a persistent performance improvement.

+++ src/main/java/com/soeguet/model/MessageTypes.java
@@ -2,13 +2,13 @@ package com.soeguet.model;
 
 public interface MessageTypes {
 
-    int DELETED = 127;
+    byte DELETED = 127;
 
-    int NORMAL = 0;
-    int LINK = 16;
-    int INTERRUPT = 25;
+   byte NORMAL = 0;
+   byte LINK = 16;
+   byte INTERRUPT = 25;
 
-    int INTERACTED = 2;
+   byte INTERACTED = 2;
 
-    int EDITED = 7;
+   byte EDITED = 7;
 }
